+++ src/main/kotlin/graphql/dataTransferObjects/InvalidBroadcast.kt
@@ -2,6 +2,6 @@ package com.neelkamath.omniChatBackend.graphql.dataTransferObjects
 
 import com.neelkamath.omniChatBackend.graphql.routing.Placeholder
 
-object InvalidBroadcast : CreateTextMessageResult {
+object InvalidBroadcast : CreateTextMessageResult, ForwardMessageResult {
     fun getPlaceholder() = Placeholder
 }