+++ docs/CHANGELOG.md
@@ -18,8 +18,8 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 
 ### Changed
 
-- Add `type InvalidBroadcast` to `union CreateTextMessageResult`.
-- Return a `type InvalidBroadcast` instead of an authorization error in `Mutation.createTextMessage`.
+- Add `type InvalidBroadcast` to `union CreateTextMessageResult` and `union ForwardMessageResult`.
+- Return a `type InvalidBroadcast` instead of an authorization error in `Mutation.createTextMessage` and `Mutation.forwardMessageResult`.
 - Return a `union ReadMessageResult` instead of a `interface Message` from `Query.readMessage`.
 - Rename the `title` field to `question` in `type Poll` and `input PollInput`.
 - Rename `Mutation.setInvitability` to `Mutation.setPublicity`.